Properly namespace the main object
Having the root namespace as the main class means we have to apply its
superclass throughout the entire namespace. This fixes that by moving
the main object to `OptionalQueryCache::QueryCache`.

This caused building/running the repository on CI services to fail
because they couldn't `bundle install` without parsing the `gemspec`,
which relied on parsing `OptionalQueryCache::Version`, which in turn
depended on `active_record`… which hadn't been installed because
`bundle install` hadn't yet been run.